added config file for constants. moved reusable thread and pynput listener functions to utils file. Moved keylogger client code to Keylogger/client. Renamed Keylogger server code to Keylogger/server.
