+++ .gitignore
@@ -1,1 +1,7 @@
+# .gitignore
+
+# Ignore compiled Python files
 __pycache__/
+*.pyc
+*.pyo
+*.pyd
