+++ config.py
@@ -1,0 +1,2 @@
+HOST = 'localhost'
+PORT = 12346
